refactor(server): fold websocket tests back inline via module split
The extracted src/websocket/tests.rs created a new src/ test home,
which tests/test_placement.rs forbids, and inlining all test lines into
websocket.rs crosses the 800-line U-16 baseline. Split the module along
its natural seams instead:

- websocket/mod.rs: ws_handler (origin + bearer auth upgrade) plus the
  full inline integration-test module
- websocket/origin.rs: origin validation helpers with their unit tests
- websocket/connection.rs: per-connection pump (bounded outbound queue,
  notification forwarding, heartbeat, shutdown)

No behavior change; every file stays under the line limit.
